rental_data_management/backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Cors
pyodbc
python-dotenv
bcrypt
pandas
openpyxl
pymysql
cryptography